refactor: tidy proxy test helpers and split worker ops contracts (#676 E+G+H)
Three small items from #676 bundled into one PR:

E. Move StoresFromMock + the structural composite interface out of
   internal/proxy/stores.go. The function was test-only API that was
   sitting in production source. New layout:
     - internal/proxy/proxytest:  AllRoles + StoresFromMock for
                                  cross-package tests
     - internal/proxy/stores_test.go:  testStoresFromMock + allRoles for
                                       proxy's own in-package tests
                                       (proxy can't import its own
                                       proxytest subpackage without a
                                       cycle).
   ~16 test files updated to call proxytest.StoresFromMock or
   testStoresFromMock as appropriate.

G. Split internal/worker/ops.go into ops_store.go (per-worker store role
   contracts) and ops_runtime.go (BackendAccess / AdmissionControl /
   DataMover / UsageAccessor + the composed Ops/CleanupOps/ScrubberOps).
   File-level rearrangement only - no API change, no test change.

H. Sanity-check confirmed: internal/cli/serve does not import
   internal/store/postgres or internal/store/sqlite. Engine choice
   already flows through DI exclusively (this was implicitly true after
   #680).

Build, vet, vet -tags integration, lint, and short tests all clean.
