+++ internal/proxy/rebalancer_test.go
@@ -152,7 +152,7 @@ func TestExecuteMoves_Concurrent(t *testing.T) {
 	obs := map[string]s3be.ObjectBackend{"src": src, "dest": dest}
 	mgr := NewBackendManager(&BackendManagerConfig{
 		Backends:        obs,
-		Stores:          StoresFromMock(store),
+		Stores:          testStoresFromMock(store),
 		Dashboard:       store,
 		Metrics:         store,
 		Order:           []string{"src", "dest"},
@@ -205,7 +205,7 @@ func TestExecuteMoves_PartialFailure(t *testing.T) {
 	obs := map[string]s3be.ObjectBackend{"src": src, "dest": dest}
 	mgr := NewBackendManager(&BackendManagerConfig{
 		Backends:        obs,
-		Stores:          StoresFromMock(store),
+		Stores:          testStoresFromMock(store),
 		Dashboard:       store,
 		Metrics:         store,
 		Order:           []string{"src", "dest"},
@@ -239,7 +239,7 @@ func TestExecuteMoves_SequentialFallback(t *testing.T) {
 	obs := map[string]s3be.ObjectBackend{"src": src, "dest": dest}
 	mgr := NewBackendManager(&BackendManagerConfig{
 		Backends:        obs,
-		Stores:          StoresFromMock(store),
+		Stores:          testStoresFromMock(store),
 		Dashboard:       store,
 		Metrics:         store,
 		Order:           []string{"src", "dest"},
@@ -434,7 +434,7 @@ func newRebalanceManager(store *mockStore, names []string) *BackendManager {
 	}
 	return NewBackendManager(&BackendManagerConfig{
 		Backends:        backends,
-		Stores:          StoresFromMock(store),
+		Stores:          testStoresFromMock(store),
 		Dashboard:       store,
 		Metrics:         store,
 		Order:           names,
@@ -731,7 +731,7 @@ func TestExecuteOneMove_DestBackendNotFound(t *testing.T) {
 	obs := map[string]s3be.ObjectBackend{"src": src}
 	mgr := NewBackendManager(&BackendManagerConfig{
 		Backends:        obs,
-		Stores:          StoresFromMock(store),
+		Stores:          testStoresFromMock(store),
 		Dashboard:       store,
 		Metrics:         store,
 		Order:           []string{"src"},
@@ -763,7 +763,7 @@ func TestExecuteOneMove_SourceGetFails(t *testing.T) {
 	obs := map[string]s3be.ObjectBackend{"src": src, "dest": dest}
 	mgr := NewBackendManager(&BackendManagerConfig{
 		Backends:        obs,
-		Stores:          StoresFromMock(store),
+		Stores:          testStoresFromMock(store),
 		Dashboard:       store,
 		Metrics:         store,
 		Order:           []string{"src", "dest"},
@@ -796,7 +796,7 @@ func TestExecuteOneMove_DestPutFails(t *testing.T) {
 	obs := map[string]s3be.ObjectBackend{"src": src, "dest": dest}
 	mgr := NewBackendManager(&BackendManagerConfig{
 		Backends:        obs,
-		Stores:          StoresFromMock(store),
+		Stores:          testStoresFromMock(store),
 		Dashboard:       store,
 		Metrics:         store,
 		Order:           []string{"src", "dest"},
@@ -828,7 +828,7 @@ func TestExecuteOneMove_MoveLocationError_CleansUpOrphan(t *testing.T) {
 	obs := map[string]s3be.ObjectBackend{"src": src, "dest": dest}
 	mgr := NewBackendManager(&BackendManagerConfig{
 		Backends:        obs,
-		Stores:          StoresFromMock(store),
+		Stores:          testStoresFromMock(store),
 		Dashboard:       store,
 		Metrics:         store,
 		Order:           []string{"src", "dest"},
@@ -866,7 +866,7 @@ func TestExecuteOneMove_MoveLocationError_CleanupFails_EnqueuesCleanup(t *testin
 	obs := map[string]s3be.ObjectBackend{"src": src, "dest": dest}
 	mgr := NewBackendManager(&BackendManagerConfig{
 		Backends:        obs,
-		Stores:          StoresFromMock(store),
+		Stores:          testStoresFromMock(store),
 		Dashboard:       store,
 		Metrics:         store,
 		Order:           []string{"src", "dest"},
@@ -907,7 +907,7 @@ func TestExecuteOneMove_MovedSizeZero_CleansUpOrphan(t *testing.T) {
 	obs := map[string]s3be.ObjectBackend{"src": src, "dest": dest}
 	mgr := NewBackendManager(&BackendManagerConfig{
 		Backends:        obs,
-		Stores:          StoresFromMock(store),
+		Stores:          testStoresFromMock(store),
 		Dashboard:       store,
 		Metrics:         store,
 		Order:           []string{"src", "dest"},
@@ -945,7 +945,7 @@ func TestExecuteOneMove_SourceDeleteFails_EnqueuesCleanup(t *testing.T) {
 	obs := map[string]s3be.ObjectBackend{"src": src, "dest": dest}
 	mgr := NewBackendManager(&BackendManagerConfig{
 		Backends:        obs,
-		Stores:          StoresFromMock(store),
+		Stores:          testStoresFromMock(store),
 		Dashboard:       store,
 		Metrics:         store,
 		Order:           []string{"src", "dest"},
@@ -1017,7 +1017,7 @@ func TestExecuteMoves_AdmissionBlocked(t *testing.T) {
 	store := &mockStore{}
 	mgr := NewBackendManager(&BackendManagerConfig{
 		Backends:        map[string]s3be.ObjectBackend{"b1": src, "b2": dst},
-		Stores:          StoresFromMock(store),
+		Stores:          testStoresFromMock(store),
 		Dashboard:       store,
 		Metrics:         store,
 		Order:           []string{"b1", "b2"},
