+++ internal/proxy/replicator_test.go
@@ -102,7 +102,7 @@ func TestReplicate_QuotaStatsError(t *testing.T) {
 	}
 	mgr := NewBackendManager(&BackendManagerConfig{
 		Backends:        map[string]backend.ObjectBackend{"b1": b1, "b2": newMockBackend()},
-		Stores:          StoresFromMock(store),
+		Stores:          testStoresFromMock(store),
 		Dashboard:       store,
 		Metrics:         store,
 		Order:           []string{"b1", "b2"},
@@ -138,7 +138,7 @@ func TestReplicate_Success(t *testing.T) {
 	}
 	mgr := NewBackendManager(&BackendManagerConfig{
 		Backends:        map[string]backend.ObjectBackend{"b1": b1, "b2": b2},
-		Stores:          StoresFromMock(store),
+		Stores:          testStoresFromMock(store),
 		Dashboard:       store,
 		Metrics:         store,
 		Order:           []string{"b1", "b2"},
@@ -172,7 +172,7 @@ func TestFindReplicaTarget_ExcludesExistingCopies(t *testing.T) {
 	store := &mockStore{getBackendFromEligible: true}
 	mgr := NewBackendManager(&BackendManagerConfig{
 		Backends:        map[string]backend.ObjectBackend{"b1": newMockBackend(), "b2": newMockBackend(), "b3": newMockBackend()},
-		Stores:          StoresFromMock(store),
+		Stores:          testStoresFromMock(store),
 		Dashboard:       store,
 		Metrics:         store,
 		Order:           []string{"b1", "b2", "b3"},
@@ -199,7 +199,7 @@ func TestFindReplicaTarget_SkipsFullBackends(t *testing.T) {
 	store := &mockStore{}
 	mgr := NewBackendManager(&BackendManagerConfig{
 		Backends:        map[string]backend.ObjectBackend{"b1": newMockBackend(), "b2": newMockBackend()},
-		Stores:          StoresFromMock(store),
+		Stores:          testStoresFromMock(store),
 		Dashboard:       store,
 		Metrics:         store,
 		Order:           []string{"b1", "b2"},
@@ -225,7 +225,7 @@ func TestSelectReplicaTarget_NoSpaceAvailable(t *testing.T) {
 	store := &mockStore{getBackendErr: core.ErrNoSpaceAvailable}
 	mgr := NewBackendManager(&BackendManagerConfig{
 		Backends:        map[string]backend.ObjectBackend{"b1": newMockBackend(), "b2": newMockBackend()},
-		Stores:          StoresFromMock(store),
+		Stores:          testStoresFromMock(store),
 		Dashboard:       store,
 		Metrics:         store,
 		Order:           []string{"b1", "b2"},
@@ -263,7 +263,7 @@ func TestCopyToReplica_Success(t *testing.T) {
 
 	mgr := NewBackendManager(&BackendManagerConfig{
 		Backends:        map[string]backend.ObjectBackend{"b1": b1, "b2": b2},
-		Stores:          StoresFromMock(&mockStore{}),
+		Stores:          testStoresFromMock(&mockStore{}),
 		Order:           []string{"b1", "b2"},
 		CacheTTL:        5 * time.Second,
 		BackendTimeout:  30 * time.Second,
@@ -293,7 +293,7 @@ func TestCopyToReplica_FailoverToSecondCopy(t *testing.T) {
 
 	mgr := NewBackendManager(&BackendManagerConfig{
 		Backends:        map[string]backend.ObjectBackend{"b1": b1, "b2": b2, "b3": b3},
-		Stores:          StoresFromMock(&mockStore{}),
+		Stores:          testStoresFromMock(&mockStore{}),
 		Order:           []string{"b1", "b2", "b3"},
 		CacheTTL:        5 * time.Second,
 		BackendTimeout:  30 * time.Second,
@@ -324,7 +324,7 @@ func TestCopyToReplica_AllSourcesFail(t *testing.T) {
 
 	mgr := NewBackendManager(&BackendManagerConfig{
 		Backends:        map[string]backend.ObjectBackend{"b1": b1, "b2": b2},
-		Stores:          StoresFromMock(&mockStore{}),
+		Stores:          testStoresFromMock(&mockStore{}),
 		Order:           []string{"b1", "b2"},
 		CacheTTL:        5 * time.Second,
 		BackendTimeout:  30 * time.Second,
@@ -404,7 +404,7 @@ func TestReplicate_RecordReplicaFails_CleansUpOrphan(t *testing.T) {
 	}
 	mgr := NewBackendManager(&BackendManagerConfig{
 		Backends:        map[string]backend.ObjectBackend{"b1": b1, "b2": b2},
-		Stores:          StoresFromMock(store),
+		Stores:          testStoresFromMock(store),
 		Dashboard:       store,
 		Metrics:         store,
 		Order:           []string{"b1", "b2"},
@@ -452,7 +452,7 @@ func TestCopyToReplica_TargetWriteFails(t *testing.T) {
 
 	mgr := NewBackendManager(&BackendManagerConfig{
 		Backends:        map[string]backend.ObjectBackend{"b1": b1, "b2": b2},
-		Stores:          StoresFromMock(&mockStore{}),
+		Stores:          testStoresFromMock(&mockStore{}),
 		Order:           []string{"b1", "b2"},
 		CacheTTL:        5 * time.Second,
 		BackendTimeout:  30 * time.Second,
@@ -483,7 +483,7 @@ func TestReplicateObject_NoTargetAvailable(t *testing.T) {
 	}
 	mgr := NewBackendManager(&BackendManagerConfig{
 		Backends:        map[string]backend.ObjectBackend{"b1": b1},
-		Stores:          StoresFromMock(store),
+		Stores:          testStoresFromMock(store),
 		Dashboard:       store,
 		Metrics:         store,
 		Order:           []string{"b1"},
@@ -522,7 +522,7 @@ func TestReplicate_SourceGoneDuringReplication(t *testing.T) {
 	}
 	mgr := NewBackendManager(&BackendManagerConfig{
 		Backends:        map[string]backend.ObjectBackend{"b1": b1, "b2": b2},
-		Stores:          StoresFromMock(store),
+		Stores:          testStoresFromMock(store),
 		Dashboard:       store,
 		Metrics:         store,
 		Order:           []string{"b1", "b2"},
@@ -583,7 +583,7 @@ func TestReplicate_HealthAware_SkipsUnhealthyTarget(t *testing.T) {
 	}
 	mgr := NewBackendManager(&BackendManagerConfig{
 		Backends:        map[string]backend.ObjectBackend{"b1": b1, "b2": cbb2, "b3": b3},
-		Stores:          StoresFromMock(store),
+		Stores:          testStoresFromMock(store),
 		Dashboard:       store,
 		Metrics:         store,
 		Order:           []string{"b1", "b2", "b3"},
@@ -637,7 +637,7 @@ func TestReplicate_HealthAware_PrefersHealthySource(t *testing.T) {
 	}
 	mgr := NewBackendManager(&BackendManagerConfig{
 		Backends:        map[string]backend.ObjectBackend{"b1": cbb1, "b2": b2, "b3": b3},
-		Stores:          StoresFromMock(store),
+		Stores:          testStoresFromMock(store),
 		Dashboard:       store,
 		Metrics:         store,
 		Order:           []string{"b1", "b2", "b3"},
@@ -681,7 +681,7 @@ func TestReplicate_HealthAware_BelowThreshold(t *testing.T) {
 	}
 	mgr := NewBackendManager(&BackendManagerConfig{
 		Backends:        map[string]backend.ObjectBackend{"b1": b1, "b2": cbb2},
-		Stores:          StoresFromMock(store),
+		Stores:          testStoresFromMock(store),
 		Dashboard:       store,
 		Metrics:         store,
 		Order:           []string{"b1", "b2"},
@@ -736,7 +736,7 @@ func TestIsBackendHealthy_CBHealthy(t *testing.T) {
 	cbb := backend.NewCircuitBreakerBackend(newMockBackend(), "b1", 3, time.Minute)
 	mgr := NewBackendManager(&BackendManagerConfig{
 		Backends:        map[string]backend.ObjectBackend{"b1": cbb},
-		Stores:          StoresFromMock(&mockStore{}),
+		Stores:          testStoresFromMock(&mockStore{}),
 		Order:           []string{"b1"},
 		CacheTTL:        5 * time.Second,
 		RoutingStrategy: config.RoutingPack,
@@ -751,7 +751,7 @@ func TestIsBackendHealthy_CBUnhealthy(t *testing.T) {
 	cbb := newTrippedCBBackend(newMockBackend(), "b1")
 	mgr := NewBackendManager(&BackendManagerConfig{
 		Backends:        map[string]backend.ObjectBackend{"b1": cbb},
-		Stores:          StoresFromMock(&mockStore{}),
+		Stores:          testStoresFromMock(&mockStore{}),
 		Order:           []string{"b1"},
 		CacheTTL:        5 * time.Second,
 		RoutingStrategy: config.RoutingPack,
