+++ internal/proxy/stores.go
@@ -57,42 +57,3 @@ type overReplicationStore struct {
 	core.QuotaStore
 }
 
-// StoresFromMock returns a Stores bag where every field points at the same
-// value. Intended for tests: pass a single mock that satisfies all role
-// interfaces (e.g. testutil.MockStore) and receive a fully-populated Stores
-// ready to drop into BackendManagerConfig.
-func StoresFromMock(m allRoles) Stores {
-	return Stores{
-		Object:           m,
-		Quota:            m,
-		Multipart:        m,
-		Replication:      m,
-		Cleanup:          m,
-		Pending:          m,
-		Integrity:        m,
-		Lifecycle:        m,
-		BackendLifecycle: m,
-		Dashboard:        m,
-		Usage:            m,
-		Lock:             m,
-	}
-}
-
-// allRoles is the structural requirement StoresFromMock imposes on its
-// argument — any value that satisfies every narrow role interface can be
-// spread across a Stores bag. Declared inline rather than exported because
-// only tests need this composite shape.
-type allRoles interface {
-	core.ObjectStore
-	core.QuotaStore
-	core.MultipartStore
-	core.ReplicationStore
-	core.CleanupStore
-	core.PendingStore
-	core.IntegrityStore
-	core.ExpiredObjectsLister
-	core.BackendLifecycleStore
-	core.DashboardStore
-	core.UsageFlusher
-	core.AdvisoryLocker
-}
